storage/storage_requirements.txt
confluent-kafka
numpy
opencv-python
requests
faiss-cpu
